requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --allow-unsafe
#
click==8.0.3
    # via typer
dnspython==2.2.0
    # via email-validator
email-validator==1.1.3
    # via yaml2rss (setup.py)
idna==3.3
    # via email-validator
jinja2==3.0.3
    # via yaml2rss (setup.py)
markupsafe==2.0.1
    # via jinja2
pydantic==1.9.0
    # via yaml2rss (setup.py)
python-magic==0.4.24
    # via yaml2rss (setup.py)
pyyaml==6.0
    # via yaml2rss (setup.py)
typer==0.4.0
    # via yaml2rss (setup.py)
typing-extensions==4.0.1
    # via pydantic
